digitrecognizer/requirements.txt
Flask==0.12.1
numpy
Pillow
requests
